pyproject.toml
[project]
name = "qdrant-stand"
version = "0.1.0"
description = "Read-only Qdrant учебный стенд с предзапечённой коллекцией для RAG-практик"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langchain==1.2.15",
    "langchain-openai==1.1.12",
    "langchain-qdrant==1.1.0",
    "langchain-text-splitters==1.1.2",
    "loguru==0.7.3",
    "pydantic==2.12.5",
    "pyyaml==6.0.3",
    "qdrant-client>=1.12,<2.0",
]

[tool.uv]
package = false
